Correct calculation of enrich score
In the current version, the if-statements don't seem to fit for me.
First, I think instead of checking the target section n2, it looks like n3 should be checked when I look at what is calculated in each case.
Second, checking the length of n, which is calculated above as the length of x doesn't make sense since this will always give an integer > 0 and therefore will always be true.
In fact, since x1, x2 and x3 are set to an empty matrix anyway in the normalizeToMatrix function you can also instead of correcting the points above just always calculate the whole formula and if up/downstream or target vector is empty it will be zero in the sum and thus not change the result.